requirements.txt
Django==4.0.4
django-extensions==3.1.5
django-tables2==2.4.1
djangorestframework==3.13.1
